Fix coverart property getter logic
Since #375 was merged, podcast cover art does not display. This change
`GLib.File.new_for_uri` for both `remote_art_uri` and `local_art_uri`
and consolidates the getter logic to eliminate null and empty check
inconsistencies. The core of this issue appears to be the the use of
`new_for_path`, which expects an absolute or relative filepath rather
than a `file://` URI.